Changed ndarray implementation to Multiarray

+++ src/modules/utils.rs
@@ -1,9 +1,8 @@
 use std::env;
-use std::vec::Vec;
 
-pub fn get_int_in_eniroment(input: &str) -> u32 {
+pub fn get_int_in_eniroment(input: &str) -> usize {
   match env::var(input) {
-    Ok(val) => val.to_string().parse::<u32>().unwrap(),
+    Ok(val) => val.to_string().parse::<usize>().unwrap(),
     Err(e) => panic!(e),
   }
 }
@@ -15,7 +14,7 @@ pub fn get_float_in_eniroment(input: &str) -> f64 {
   }
 }
 
-pub fn get_axis_size() -> Vec<u32> {
+pub fn get_axis_size() -> Vec<usize> {
   let mut axis_size = Vec::new();
   axis_size.push(get_int_in_eniroment("FIRST_AXIS"));
   axis_size.push(get_int_in_eniroment("SECOND_AXIS"));
